+++ src/main/java/faang/school/notificationservice/config/context/RedisConfig.java
@@ -1,58 +1,58 @@
-package faang.school.notificationservice.config.context;
-
-import faang.school.notificationservice.listener.CommentEventListener;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
-import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.listener.ChannelTopic;
-import org.springframework.data.redis.listener.RedisMessageListenerContainer;
-import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
-
-@Configuration
-@RequiredArgsConstructor
-public class RedisConfig {
-
-    @Value("${spring.data.redis.host}")
-    private String host;
-    @Value("${spring.data.redis.port}")
-    private int port;
-    @Value("${spring.data.redis.channel.comment")
-    private String commentChannel;
-
-    @Bean
-    public JedisConnectionFactory jedisConnectionFactory() {
-        RedisStandaloneConfiguration configuration = new RedisStandaloneConfiguration(host, port);
-        return new JedisConnectionFactory(configuration);
-    }
-
-    @Bean
-    public RedisTemplate<String, Object> redisTemplate() {
-        final RedisTemplate<String, Object> template = new RedisTemplate<>();
-        template.setConnectionFactory(jedisConnectionFactory());
-        return template;
-    }
-
-
-    @Bean
-    public MessageListenerAdapter commentEventListenerAdapter(CommentEventListener commentEventListener) {
-        return new MessageListenerAdapter(commentEventListener);
-    }
-
-    @Bean
-    public ChannelTopic commentEventTopic() {
-        return new ChannelTopic(commentChannel);
-    }
-
-    @Bean
-    public RedisMessageListenerContainer redisContainer(MessageListenerAdapter commentEventListener) {
-        RedisMessageListenerContainer container = new RedisMessageListenerContainer();
-        container.setConnectionFactory(jedisConnectionFactory());
-
-        container.addMessageListener(commentEventListener, commentEventTopic());
-        return container;
-    }
-}
+//package faang.school.notificationservice.config.context;
+//
+//import faang.school.notificationservice.listener.CommentEventListener;
+//import lombok.RequiredArgsConstructor;
+//import org.springframework.beans.factory.annotation.Value;
+//import org.springframework.context.annotation.Bean;
+//import org.springframework.context.annotation.Configuration;
+//import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
+//import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
+//import org.springframework.data.redis.core.RedisTemplate;
+//import org.springframework.data.redis.listener.ChannelTopic;
+//import org.springframework.data.redis.listener.RedisMessageListenerContainer;
+//import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
+//
+//@Configuration
+//@RequiredArgsConstructor
+//public class RedisConfig {
+//
+//    @Value("${spring.data.redis.host}")
+//    private String host;
+//    @Value("${spring.data.redis.port}")
+//    private int port;
+//    @Value("${spring.data.redis.channel.comment")
+//    private String commentChannel;
+//
+//    @Bean
+//    public JedisConnectionFactory jedisConnectionFactory() {
+//        RedisStandaloneConfiguration configuration = new RedisStandaloneConfiguration(host, port);
+//        return new JedisConnectionFactory(configuration);
+//    }
+//
+//    @Bean
+//    public RedisTemplate<String, Object> redisTemplate() {
+//        final RedisTemplate<String, Object> template = new RedisTemplate<>();
+//        template.setConnectionFactory(jedisConnectionFactory());
+//        return template;
+//    }
+//
+//
+//    @Bean
+//    public MessageListenerAdapter commentEventListenerAdapter(CommentEventListener commentEventListener) {
+//        return new MessageListenerAdapter(commentEventListener);
+//    }
+//
+//    @Bean
+//    public ChannelTopic commentEventTopic() {
+//        return new ChannelTopic(commentChannel);
+//    }
+//
+//    @Bean
+//    public RedisMessageListenerContainer redisContainer(MessageListenerAdapter commentEventListener) {
+//        RedisMessageListenerContainer container = new RedisMessageListenerContainer();
+//        container.setConnectionFactory(jedisConnectionFactory());
+//
+//        container.addMessageListener(commentEventListener, commentEventTopic());
+//        return container;
+//    }
+//}
