+++ src/main/java/faang/school/notificationservice/listener/CommentEventListener.java
@@ -2,11 +2,13 @@ package faang.school.notificationservice.listener;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import faang.school.notificationservice.client.UserServiceClient;
+import faang.school.notificationservice.config.context.UserContext;
 import faang.school.notificationservice.dto.CommentEventDto;
 import faang.school.notificationservice.dto.UserDto;
-import faang.school.notificationservice.message_builder.CommentMessageBuilder;
 import faang.school.notificationservice.message_builder.MessageBuilder;
 import faang.school.notificationservice.service.NotificationService;
+import jakarta.validation.constraints.NotNull;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.connection.Message;
 import org.springframework.stereotype.Component;
 
@@ -16,21 +18,21 @@ import java.util.Locale;
 @Component
 public class CommentEventListener extends AbstractEventListener<CommentEventDto> {
 
-    UserServiceClient userServiceClient;
-
-    public CommentEventListener(UserServiceClient userServiceClient,
+    @Autowired
+    public CommentEventListener(ObjectMapper objectMapper,
+                                UserServiceClient userServiceClient,
                                 List<MessageBuilder<CommentEventDto>> messageBuilders,
-                                List<NotificationService> notificationService,
-                                ObjectMapper mapper) {
-        super(userServiceClient, messageBuilders, notificationService, mapper);
-        this.userServiceClient = userServiceClient;
+                                List<NotificationService> notificationServices,
+                                UserContext userContext) {
+        super(objectMapper, userServiceClient, messageBuilders, notificationServices, userContext);
     }
 
-    public void onMessage(Message message, byte[] pattern) {
-        CommentEventDto event = getEvent(message, CommentEventDto.class);
-        String notifyMessage = getMessage(event, Locale.ENGLISH);
-
-        UserDto user = userServiceClient.getUser(event.getAuthorPostId());
-        sendNotification(user, notifyMessage);
+    @Override
+    public void onMessage(@NotNull Message message, byte[] pattern) {
+        handleEvent(message, CommentEventDto.class, event -> {
+            String text = getMessage(event, Locale.ENGLISH);
+            UserDto user = userServiceClient.getUser(event.getAuthorPostId());
+            sendNotification(user, text);
+        });
     }
 }
