+++ src/main/java/faang/school/notificationservice/listener/MentorshipOfferedListener.java
@@ -5,26 +5,27 @@ import faang.school.notificationservice.client.UserServiceClient;
 import faang.school.notificationservice.config.context.UserContext;
 import faang.school.notificationservice.dto.MentorshipOfferedEvent;
 import faang.school.notificationservice.dto.UserDto;
-import faang.school.notificationservice.service.MessageBuilder;
+import faang.school.notificationservice.message_builder.MessageBuilder;
 import faang.school.notificationservice.service.NotificationService;
 import jakarta.validation.constraints.NotNull;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.connection.Message;
-import org.springframework.data.redis.connection.MessageListener;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
 import java.util.Locale;
 
-@Component
 @Slf4j
-public class MentorshipOfferedListener extends AbstractEventListener<MentorshipOfferedEvent> implements MessageListener {
+@Component
+public class MentorshipOfferedListener extends AbstractEventListener<MentorshipOfferedEvent> {
 
     @Autowired
-    public MentorshipOfferedListener(ObjectMapper objectMapper, UserServiceClient userServiceClient,
+    public MentorshipOfferedListener(ObjectMapper objectMapper,
+                                     UserServiceClient userServiceClient,
                                      List<MessageBuilder<MentorshipOfferedEvent>> messageBuilders,
-                                     List<NotificationService> notificationServices, UserContext userContext) {
+                                     List<NotificationService> notificationServices,
+                                     UserContext userContext) {
         super(objectMapper, userServiceClient, messageBuilders, notificationServices, userContext);
     }
