+++ src/main/java/faang/school/notificationservice/message_builder/MessageBuilder.java
@@ -1,11 +1,10 @@
 package faang.school.notificationservice.message_builder;
 
-import org.springframework.stereotype.Component;
-
 import java.util.Locale;
 
-@Component
 public interface MessageBuilder<T> {
-    String getMessage(Locale locale, T Event);
-    Class<?> getBuilderType();
+
+    String buildMessage(T eventType, Locale locale);
+
+    Class<?> supportsEventType();
 }
