--- src/main/java/faang/school/notificationservice/service/MessageBuilder.java
@@ -1,10 +1,0 @@
-package faang.school.notificationservice.service;
-
-import java.util.Locale;
-
-public interface MessageBuilder<T> {
-
-    String buildMessage(T eventType, Locale locale);
-
-    Class<?> supportsEventType();
-}
