+++ src/main/java/faang/school/notificationservice/service/email/EmailService.java
@@ -3,6 +3,7 @@ package faang.school.notificationservice.service.email;
 import faang.school.notificationservice.dto.UserDto;
 import faang.school.notificationservice.service.NotificationService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
@@ -12,19 +13,22 @@ import org.springframework.stereotype.Service;
 public class EmailService implements NotificationService {
 
     private final JavaMailSender mailSender;
-
-    @Override
-    public UserDto.PreferredContact getPreferredContact() {
-        return UserDto.PreferredContact.EMAIL;
-    }
+    @Value("${spring.mail.username}")
+    private String senderMail;
 
     @Override
     public void send(UserDto user, String text) {
         SimpleMailMessage message = new SimpleMailMessage();
+        message.setFrom(senderMail);
         message.setTo(user.getEmail());
         message.setSubject("CorporationX");
         message.setText(text);
         mailSender.send(message);
+    }
+
+    @Override
+    public UserDto.PreferredContact getPreferredContact() {
+        return UserDto.PreferredContact.EMAIL;
     }
 
 }
